refactor: Deathmason death event into onDeathEvent
Fixes #355
